// apps/core-service/src/app/modules/core/core-auth/dto/core-auth.dto.ts
//=============================================================================
//Licensed Materials - Property of Bambee
//(C) Copyright Bambee 2022
//All Rights Reserved
//=============================================================================

export interface CoreAuthDTO {
  readonly email: string;
  readonly id: string;
  readonly username: string;
  readonly tosAcceptedAt?: Date;
}